% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ee_diffex_functions.R
\name{get_diffexed_genes}
\alias{get_diffexed_genes}
\title{get_diffexed_genes}
\usage{
get_diffexed_genes(fit, gene_column = "symbol", diffex_type = c("any_probe",
  "all_probes"), ...)
}
\arguments{
\item{fit}{A limma fit object (class: MArrayLM) as returned
by eBayes. If the initial model/analysis contained more than one contrast
the user should select a single contrast prior to passing \code{fit} into
\code{get_diffexed_genes}, by doing fit[, some.column] - if info on more
than one contrast is present in the \code{fit} object, the function throws
an error.}

\item{gene_column}{The name of the column within \code{fit$genes}
that contains the gene-symbols or gene-ids that are to be returned by
this function. For example, if you want to receive a subset of the
"entrez.id" column of \code{fit$genes} then specify
\code{gene_column = "entrez.id"}. The function dies if the specified
column is absent from the \code{fit$genes} dataframe.}

\item{diffex_type}{The user selects either \code{"any_probe"} or
\code{"all_probes"} to indicate the type of analysis to be performed.}

\item{...}{Further arguments that are to be passed through
to limma::decideTests. For example, if you want to set the \code{lfc}
filter to 1 (in decideTests; corresponding to keeping probes only if they
are at least 2-fold up/down-regulated), you can specify it as follows:
\code{get_diffexed_genes(some.fit, some.column, some.type, lfc = 1)}.}
}
\value{
A subvector of the inidicated 'gene_column' in fit fit$genes
  annotation dataframe.
}
\description{
Extract genes from a limma-fit object (resulting from an eBayes call) that
are differentially expressed. Two different methods are used for calling
differential expression:
  \code{any_probe} - at least one probe for that gene must be differentially
  expressed.
  \code{all_probes} - every probe for that gene must be differentially
  expressed (and must be differentially expressed in the same direction).
This doesn't return information about the direction of change, and only
returns a vector of gene symbols or ids, as extracted from the \code{genes}
dataframe in the limma-fit object \code{fit}.
}

